feat: call GC.collect to free up unused memory
If the uploader, at a moment of time uses a lot of memory and no
activity is being performed on the uploader (like file retrieval,
uploading, whatever), the GC is not called, since it only APPEARS to be
called after a certain amount of function calls. Don't quote me on this,
but that is what I feel. I maybe should ask crystal devs lol