fix: Use inline script to remove form and use noscript to show upload history warning
To make the a smooth transition from the noJS upload form to the JS one, an inline script was needed so the browser can take it and parse it with the html page. This fixes that small time gap when waiting for the upload javascript to load and hidding the noJS upload form.

TODO: Add nonce or checksum for the inline script since it's considered a security risk.
https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Headers/Content-Security-Policy/script-src